throw `NotFoundException` in `deleteBeerById` if beer not found, add corresponding integration tests in `BeerControllerIT`

+++ src/main/java/dev/dead/spring6restmvc/controllers/BeerController.java
@@ -33,6 +33,9 @@ public class BeerController {
     @DeleteMapping(BEER_ID_URL)
     public @NotNull ResponseEntity deleteBeerById(@PathVariable("beerId") UUID beerId) {
         log.debug("Delete Beer - Controller id: {}", beerId);
+        if (!beerService.deleteBeerById(beerId)) {
+            throw new NotFoundException();
+        }
         beerService.deleteBeerById(beerId);
         return ResponseEntity.noContent()
                 .build();
